+++ src/tests/ftest/security/pool_security_acl.yaml
@@ -4,17 +4,19 @@ hosts:
   test_servers:
     - server-A
   test_clients:
-    - client-A   #server client same host
+    - server-B   #server client different host
   transport_config:
     - allow_insecure: False
     - ca_cert: ~/.daos/daosCA.crt
     - cert: ~/.daos/daos_agent.crt
     - key: ~/.daos/daos_agent.key
     - server_name: server
-timeout: 1200
+timeout: 200
 server_config:
   name: daos_server
   port: 10001
+server_manager:
+  recreate: true
 daos_server:
   start:
      insecure: False
